Support HQT on VLLM - KVCache and Mark Step uses

+++ vllm/utils.py
@@ -31,7 +31,7 @@ STR_DTYPE_TO_TORCH_DTYPE = {
     "half": torch.half,
     "bfloat16": torch.bfloat16,
     "float": torch.float,
-    "fp8": torch.uint8,
+    "fp8": torch.float8_e4m3fn,
 }
 
 
@@ -68,6 +68,7 @@ class Softmax(torch.nn.Module):
 
       def forward(self, x, dim = None, inv_head = None):
         return torch.softmax(x, dim)
+
 
 class LRUCache(Generic[T]):
